+++ src/Network/ExpressRouteLag.Autorest/test/New-AzExpressRouteLag.Tests.ps1
@@ -1,0 +1,31 @@
+if(($null -eq $TestName) -or ($TestName -contains 'New-AzExpressRouteLag'))
+{
+  $loadEnvPath = Join-Path $PSScriptRoot 'loadEnv.ps1'
+  if (-Not (Test-Path -Path $loadEnvPath)) {
+      $loadEnvPath = Join-Path $PSScriptRoot '..\loadEnv.ps1'
+  }
+  . ($loadEnvPath)
+  $TestRecordingFile = Join-Path $PSScriptRoot 'New-AzExpressRouteLag.Recording.json'
+  $currentPath = $PSScriptRoot
+  while(-not $mockingPath) {
+      $mockingPath = Get-ChildItem -Path $currentPath -Recurse -Include 'HttpPipelineMocking.ps1' -File
+      $currentPath = Split-Path -Path $currentPath -Parent
+  }
+  . ($mockingPath | Select-Object -First 1).FullName
+}
+
+Describe 'New-AzExpressRouteLag' {
+    It 'CreateExpanded' {
+        $lag = New-AzExpressRouteLag -Name $env.expressRouteLagName -ResourceGroupName $env.resourceGroupName -Location $env.location -PeeringLocation $env.peeringLocation -BandwidthInGbps $env.bandwidthInGbps -Encapsulation $env.encapType -BillingType $env.billingType -NumberOfPort $env.noOfPorts -MinimumActivePortsRequired $env.minLinks -LacpTimer $env.lacpTimer
+        $lag.Name | Should -Be $env.expressRouteLagName
+        $lag.ProvisioningState | Should -Not -BeNullOrEmpty
+    }
+
+    It 'CreateViaJsonFilePath' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+
+    It 'CreateViaJsonString' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+}