+++ src/Network/ExpressRouteLag.Autorest/test/New-AzExpressRouteLagLoa.Tests.ps1
@@ -1,0 +1,43 @@
+if(($null -eq $TestName) -or ($TestName -contains 'New-AzExpressRouteLagLOA'))
+{
+  $loadEnvPath = Join-Path $PSScriptRoot 'loadEnv.ps1'
+  if (-Not (Test-Path -Path $loadEnvPath)) {
+      $loadEnvPath = Join-Path $PSScriptRoot '..\loadEnv.ps1'
+  }
+  . ($loadEnvPath)
+  $TestRecordingFile = Join-Path $PSScriptRoot 'New-AzExpressRouteLagLOA.Recording.json'
+  $currentPath = $PSScriptRoot
+  while(-not $mockingPath) {
+      $mockingPath = Get-ChildItem -Path $currentPath -Recurse -Include 'HttpPipelineMocking.ps1' -File
+      $currentPath = Split-Path -Path $currentPath -Parent
+  }
+  . ($mockingPath | Select-Object -First 1).FullName
+}
+
+Describe 'New-AzExpressRouteLagLOA' {
+    It 'GenerateExpanded' {
+        $loa = New-AzExpressRouteLagLOA -ExpressRouteLagName $env.expressRouteLagName -ResourceGroupName $env.resourceGroupName -CustomerName 'Contoso' -Member @('member1')
+        $loa | Should -Not -BeNullOrEmpty
+        $loa.EncodedContent | Should -Not -BeNullOrEmpty
+    }
+
+    It 'GenerateViaJsonString' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+
+    It 'GenerateViaJsonFilePath' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+
+    It 'Generate' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+
+    It 'GenerateViaIdentityExpanded' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+
+    It 'GenerateViaIdentity' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+}